kgo.Client: avoid panic in OffsetFetchRequest when coordinator is not loaded
Internally, OffsetFetchRequest is sharded, but only on old brokers. We
first try not sharding. To have this one-or-the-other behavior, we wrap
the request with an internal pinned type before issuing it. Before we
return the response shards to the user, we remove any pin wrapping.

The offset fetch sharder accidentally did not return a pinned req in
cases that were known to be failures (i.e. were actually not going to be
issued). This wasn't caught in testing because the only time this really
happens is when:

* the FindCoordinator request outright fails
* the broker's coordinator is not yet loaded

The latter case happens for a brief window when using groups for the
first time in a new cluster.

Rather than always return a pinned req, we instead check if the request
actually is pinned. This is more bullet proof for the future and
actually eliminates the need for the current helper "is this pinned"
function.

This also adds a tiny integration test that theoretically should trigger
the issue occasionally -- if this test runs first in a test suite, then
it will panic. I've also tested this locally before and after this
change on a new cluster both times.

Lastly, this removes accidentally requesting an empty group in
FindCoordinators if the request was batched.