pkg/kmsg: add Record.TimestampDelta64
This is a different approach to the prior commits, rather than bumping
to a v2 kmsg which also requires v2 kgo, we add TimestampDelta64 and
manually use it rather than the old field.

To ensure backwards compatibility, we use the old TimestampDelta if the
new 64 one is 0, and we decode into both.