kgo: patch ConsumePartitions regression from 1.11.5
AddConsumeTopics was patched in 1.11.5 to actually work for the direct
consumer. Unfortunately, this broke ConsumePartitions by always
consuming all partitions. The prior commits introduced tests ensuring
we do not regress on either of these issues. This commit introduces the
fix.

This also fixes two other small things:
* We retry more on ILLEGAL_GENERATION while committing, which was seen
once in tests (on GH, slow). We also sleep 10ms before retrying, which
is short but hopefully enough of a pause when retrying.
* We change blockingMetadataFn to wait for the function if the function
is called. This was noticed in the TestDirectPartitionPurge test.