kgo: always strip retryable errors when consuming
UnknownTopicID is now returned frequently on Kafka 3.4, and it looks due
to topic state not being synced immediately on all brokers. This was
causing TestIssue337 to fail due to an unexpected error in the fetch
responses.

Now, we just strip all retryable errors.

We also improve the test a bit.