kversion: cut Kafka 3.4
Tested with kcl, we can now probe 3.4 successfully.

This required a bit of a change to version guessing (and I really hope
KIP-885 is in by the next release...). Previously, we would skip the few
internal APIs by default. Now, if the ApiVersion response includes an
API we are skipping, we check it.