franz-go: pin to 1.18, remove 1.15 and 1.16 backcompat
Officially, 1.15 and 1.16 have not been supported for two years.
The klauspost/compress dependency has pinned to 1.18 for half a year,
so I _suspect_ most people using this library are actually on 1.18+.
Removing support for pre-1.18 allows us to remove isNetClosedErr
files (go115 and go116) and allows us to switch from `interface{}` to
`any`.