kgo: add ConsumeRecreatedTopics option, for UNKNOWN_TOPIC_ID
FOR CONSUMING

This is a bit of an involved change for consuming, but overall, it
works as it says on the tin.

FOR PRODUCING

We cannot have this same type of logic in the producer until produce
requests support topic IDs. As well, the producer will be more
complicated -- we will need to worry about ordering issues. So, we avoid
doing anything for producers until the day comes.

For now, if a person recreates a topic, the person will most likely
receive an OutOfOrderSequenceNumber error which looks like data loss.

If a person recreates a topic with fewer partitions, the partitions that
were deleted now have all records failed on them (and are no longer
considered writeable). The client will enter a state of permanently
failing any record that is consistently hashed to the old partition, and
permanently trying to reload metadata. The best fix is for the user to
purge the deleted topic from the client.

A more involved, more complicated, better change would be to react to
the topic ID changing within the metadata loop, but that requires some
more complicated changes such as making the topic ID atomic, and there
are race conditions that frankly aren't worth diving into at the moment
(we need to track current and last topic ID -- this was initially the
approach I was taking for consuming, but it's not needed there, but it
is needed for producing).

Closes #358.