plugin/kzap: support LogLevelNone
Previously by default, it was not possible to opt into LogLevelNone for
the "level" call.

What this would mean is that if a person had no logging, we would could
call zap.Error and then their logger would just ignore the log. Now, we
avoid calling zap.Error at all.