bugfix kgo: do not default to eager if there is any eager balancer
The previous cooperative check was wrong: if you used multiple group
balancers, if any were eager, the client would think you were eager
consuming.

However this bug did not affect all areas of group balancing, which
actually made the bug worse. If everything in the code went eager, then
the bug would just be "you saw more stop the world than was intended".
Instead, if the chosen balancer was cooperative, then cooperative
balancing would occur for eager consumers. On rebalance, partitions
would be lost, and then a cooperative rebalance would not occur, meaning
the partitions would be stuck until another rebalance.

We fix this by saving if we are cooperative based on the rebalancer that
is actually chosen.

This is an upgrade path that should happen once -- once to cooperative
-- but we can downgrade if the user adds a random eager balancer. This
is not supported per KIP-429 so we just warn log and continue as best we
can (likely with duplicate consumption).

Closes #366.