kgo: avoid counting pinReq version failures against retries
Technically, a request that goes to the sharding logic that is pinned to
a minimum version that fails due to errBrokerTooOld has never been
issued. Currently, if a request is enqueued for a long time due to the
client being throttled, it can fail with errBrokerTooOld even though it
was never issued.

Instead, if the request is pinned and the error is errBrokerTooOld, we
always retry immediately (up to 3 times, for safety).