kgo: purge missing-from-meta topics while regex consuming
While regex consuming, if a topic is deleted, we previously would
forever continue to try fetching it. Now, if a topic is missing from a
metadata response, we internally purge it.

We delete immediately with the assumption that once a topic is
discovered, there is enough broker consensus that it will not be somehow
accidentally missing on the next metadata update -- i.e., if a topic is
missing, it's definitely deleted.

We only do this for regex consuming; for non-regex, the assumption is
that if you specify topics to consume, you want them. If they are
missing in a metadata response, we will continue to try to update the
metadata even if the topic was deleted (and we will continue trying to
fetch).

Closes #434