Add the ability to unregister metrics.
We have a program that creates a new kgo client for each long runnning
request to the service and need a way to have unique metrics per long
running client/connection, but don't want the metrics to hand around
and be reported after the client is closed.

This allows us to unregister the metrics after the client is closed.