kgo: fix new niche problem against Kafka 3.5
Kafka 3.5 seems to have a problem of
(a) creating a topic
(b) replying to a client with the current topic information
(c) migrating partition leadership immediately to another broker
(d) having that other broker reply with unknown partition for a brief
window

For regex consuming, this meant the client was purging the topic because
it would first see the topic, then not see the topic.

We now allow a 15s window of retries for regex consuming where a topic
can be missing. A previous attempt at this patch used 3 tries but that
also seemingly was not working.

Updates #434.