kversion: fix version detection for Kafka v2.7 through 3.4
Kafka v3.4 added envelope support for the Envelope request to the zk
based broker, so I added that to kversion.

This made Envelope a required request for version detection ever since
it was introduced -- v2.7 for raft -- and pinned version detection to
"at least 2.7" for all cluster between 2.7 and 3.4.

We now ignore the envelope key when version guessing because ultimately
other keys are differentiating enough. This allows version detection to
now again correctly guess 2.7 through 3.4.

Tested against zk 3.1, 3.2, 3.5, as well as kraft 3.5 manually, and adds
a unit test for the 3.1 versions.

Closes #536.