pkg/sr: update for new APIs and to be forward compatible
With an HTTP API, the API can add query parameters periodically, any
addition of which breaks a Go API where all query parameters are spelled
out as method arguments.

Now, we have moved all query parameters into values in a context, and
added a helper sr.WithParams to make it easy to build parameters. All
APIs document which parameters they support. We also have global
parameters on the client itself, if you always want to show deleted or
something.

This also adds metadata and ruleset to the schema type, even though
the validation of these is really only possible in the Java client (as
documented via limitations).

Lastly, this adds support for a few missing APIs:
* SubjectsByID
* SchemaVersionsByID
* SubjectVersions
* SchemaTextByVersion
* AllSchemas

This breaks all CompatibilityLevel APIs, since the compatibility level
stuff has been expanded to a whole bunch of compatibility /
configuration settings. Namely, we just drop the "Level".

This breaks all APIs that had ShowHideDeleted arguments, favoring
instead the new ShowDeleted Param.

We keep the DeleteHow param, because it is important at call sites to
see that you are using hard deletion.