kfake: do not panic if NewCluster fails
Named returns were interfering with defer-err-cleanup. Although a
Cluster is initialized, defer close on error was running on a nil
Cluster because the return was nil, err.

We remove named returns so that the defer runs on the actual initialized
Cluster variable.

Tested by modifying main.go with the code in the issue; panics before,
does not panic after.

Closes #566.