kadm: fix occasionally empty topic/partitions in Lag
If listing end offsets errored, then the lag would have no
topic/partition. We now add Topic and Partition to the Lag itself so
that we can always ensure we set these fields to be non-empty (and this
makes it easier to reason about what field to use if you use
groupLags.Sorted()).