kgo: allow PreTxnCommitFnContext to modify empty offsets
This builds the TxnOffsetCommitRequest early so that the hook can modify it. If the modified request has no topics to commit, then we abort as though uncommitted was empty.

This moves initialization of the producerID before addOffsetsToTxn. I think that should not matter, but it's a non-trivial change in internal behavior.