kgo group: switch memberID and generation to atomics
Previously these required a mutex on write and read because of the rare
(and mostly erroneous)  chance that a person is committing during a
rebalance.

Well, that makes the next transactional commit harder and is overkill --
switching to atomics doesn't change any correctness bit but allows us to
worry about deadlocks just a bit less.