pkg/kgo consumer group: only begin autocommitting once group is managed
This takes all goroutines out of client initialization sub functions (ie
not the explicit ones _in_ client initialization) and avoids starting a
goroutine if we never even make it to the manage loop anyway (which only
happens once we finally fetch metadata)

Also skips autocommitting if there is nothing to commit. This is the
same behavior as previously, but now we have a log that explicitly says
the commit is being skipped.