kgo: fix bug
Previously, if CommitOffsetsSync was called during a rebalance, and the
context being used is canceled while the rebalance was occurring, then
the client would deadlock.

Internally, committing is blocked if a rebalance is actively happening.
There's some complex logic to have, effectively, a cancelable lock
if the user wants to not wait for a rebalance to complete while trying
to commit offsets.

There was a bug, and since the fix is one line, it's easier to see than
explain.

Closes #668.