kadm: use -1 for the commit if a topic is not committed to
Previously, if a topic had commits but some partitions in it did not,
then Lag would do the CORRECT thing and show -1 for partitions that did
not have a commit. What was NOT correct was that if an ENTIRE topic had
no commits, then Lag would show the commit as 0.