kgo: ignore aborted txns if using READ_UNCOMMITTED
Kafka does not return AbortedTransactions in the fetch response if the
client is using read uncommitted, so kgo works perfectly against Kafka.
Redpanda DOES return AbortedTransactions (today, although this is being
changed) even if the client is using READ_UNCOMMITTED, which causes
records to be skipped. Redpanda is changing, but we can address this in
kgo as well.