kgo: fix accounting when topics/partitions are {,un}paused for PollRecords
Topics or partitions that were paused (in `takeNBuffered` specifically,
which is used by PollRecords) did not have proper accounting for
stripped (removed) topics or partitions. PollFetches does not have this
problem because the logic is different -- we build a big fetch, call our
accounting logic, and then remove the stripped topics & partitions.
Instead in takeNBuffered, we skip topics and partitions as we build a
fetch and then once the fetch is built, do the accounting.

The fix is to build a separate internal-only fetch of everything that
was stripped and pass that to our accounting hooks.

Fixes #865.