kgo: send all metadata requests through the internal mapped cache
Issue #800 was created as a follow up idea to strengthen caching of
metadata requests in the client. This pushes the mapped metadata caching
logic deeper into the guts of issuing metadata requests, so that no
caching is ever missed. The next commit will introduce a new API to
request potentially cached metadata.

For #800.