kgo: fix off-by-one with retries accounting
Previously, if our tries was equal to the max retries, we would not
retry. Instead, we should try again -- the number of tries is one less
than the number of retries. We tried once, we retried 0 times.

Now, when we try twice, we have retried once, and we can fail properly.
This also means that failing once will cause a metadata reload, rather
than immediately failing the record and doing nothing.

This also switches freezing a batch to a dedicated bool field rather
than relying on a secondary effect of counting how many times a batch
has been tried. This allows us to move accounting of when a batch is
actually tried to where it is actually being written, avoiding double
counting of tries if the batch actually ends up not being sent after
being selected to be sent (due to add partitions to txn failure or
something).

Closes #937.