kversions: entirely reimplement internals
For >1yr I've known that the prior implementation wouldn't hold up into
the future. The prior implementation accurately detected zookeeper
releases, but was not entirely accurate for kraft releases and would be
more inaccurate as time went on (specifically, the prior impl had some
problems with the tons of missing keys that kraft has). Now that kraft
is the default, well, the problem needs to be addressed.

This specifically defines all versions for zookeeper, the kraft
controller, and the kraft broker separately. I ran the KRaft broker back
to 2.8 when it was feature preview to define the versions accurately,
and then tested this against kcl.

Interestingly, this surfaced that zookeeper Kafka v3.3 and 3.4 had equal
keys advertised. So, sometimes version detection can't work.

This also removes the big define-variables-on-startup-funcs that existed
before, so the version initialization happens only when you use a
kversion version function for the first time. This should remove the
kversions package from affecting program startup time at all.

Next step is to define protocols for all missing requests.