kversion: pin Stable to 3.8, kgo: document Stable safety
Until kmsg is released AND THEN kgo supports all features in 3.9 and
4.0, we need to pin kversion.Stable() to 3.8.

A person that bumps kmsg before kgo supports the new features WILL NOT
automatically opt into those new features, because kmsg & kversion are
the same package and tagged in tandem. The old kgo they are using will
return an old Stable that pins the max version to use -- below the new
kmsg max version.