kversion: return a merged version from Stable
Stable is not used for version guessing, it is used for pinning what a
kgo.Client itself can do. We want to use a merged version for Stable
so that the client *can* handle every request.

This also now opts Stable to return the latest versions, meaning kgo
needs to be updated in tandem with this PR.

Also removes a text file previously erroneously committed.