generate: enforce all tags must have defaulters; kmsg: fix bug
Tags should only be encoded if they are not the default for their type.
kmsg v1.11.0 introduced a tagged uuid field on FetchRequest, which is
meant to be set only by brokers for inter-broker communication.

The uuid type previously did not have a defaulter in the generator, and
the resulting generated code *always* encoded the ReplicaDirectorID
field. If you upgraded kmsg and used it against an old Kafka, Kafka
would fail to parse the request.

This commit now enforces that all tagged types have defaulters defined
(the generator will panic if that is not the case), adds a defaulter for
the uuid type, and fixes the generated code.

Closes #961.