kgo: fixup for KIP-890
Tests occasionally failed with an INVALID_TXN_STATE error from EndTxn.

This is because the prior AddPartitionsToTxn logic was unconditionally
swapping addedToTxn to true, and then if produce failed (causing the
partition NOT to be added to the txn), addedToTxn was not swapped back
to false. The prior code always swapped things back to false when the
AddPartitionsToTxn req failed.

Now, we move the produce version 12 check in front of addedToTxn being
set to true -- short circuiting so that addedToTxn is never set.
Instead, we add a separate conditional where partitions that produced
successfully get added to the txn.

Tests no longer fail.