kgo: expose IsRetryableBrokerErr
It is a smaller delta to expose a public func wrapping a private func
than to make the private func public and change all usages of it.

I'm switching to Retryable rather than the currently exposed
kerr.Retriable or Broker.RetriableRequest. This is an inconsistency, but
Retryable feels more correct and what I've moved to for the past years:

    https://english.stackexchange.com/a/305274

Closes #877.