kgo: cleanup Compressor.Compress interface
Compress should not expose the internals of what produce request version
is being used during compression. Instead, we replace this with compress
flags which can be extended in the future if needed. The current only
compress flag is disable gzip.

This allows the compressor to be used in other areas of the code
unrelated to producing.

The dst bytes.Buffer is a bit annoying if used outside of produce but eh
it's easy enough to just throw a new(bytes.Buffer) at it, if we do not
want to worry about the bytes.Buffer pooling.

Uint32 flags should hopefully be enough; since the implementation will
be in user hands, I don't want to go overboard on some opaque struct
with complex behavior.