kgo: observe client metrics (and remove some optional ones)
This adds a code to the few places needed to observe required (and a
_few_ optional) client metrics.

Some metrics were removed in this commit; the rebalance latency is
underspecified, the fetch latency seems to duplicate consumer node
request latency, and assigned partitions seems like medium signal and is
tedious to work in. Thus, we are left with 12 metrics, 6 of which are
required, of the 21 recommended metrics in the KIP. That's a fine start
and we can stick with it unless requested.