kgo: add ConsumeStartOffset, expand offset docs, update readme KIPs
ConsumeResetOffset has always been confusing when trying to figure out
what exactly it does and when. The main point of confusion (for me) has
always been trying to reason about where I want to start consuming new
partitions vs. where to reset to when a problem is encountered.

Well, we can split the two into dedicated options to resolve the
confusion.

This also hopefully cleans up and clarifies which offset is used when.
As it turns out, there really is only one place where the reset offset
is used, and it's used very rarely (a niche sub-case of when
OffsetOutOfRange is encountered). It's likely best to just leave it as
NewOffset().AtStart().

I've read through KIP-1106 a few times, and given that the client
already tries to reset to the best offset (right after the last
succesfully consumed offset, otherwise the end), there is not really a
reason to add duration based offset resetting to this client. I've noted
that the kip is skipped in the readme.