kgo: remove all EndAndBeginTransaction internal "optimizations"
The optimizations are fundamentally unsafe as of KIP-890.

The first difficult bug to reproduce was that `p.inflight` incrementing
needed to be *ABOVE* grabbing the producer ID in sink.go. IF it was
below, then the producer ID could be grabbed, THEN we would block to
create the request, THEN EndTxn would finish, THEN we would create a new
produce request with the old ID & epoch.

Fixing that: producers reset sequence numbers to 0 for all partitions
after EndTxn.

* First produce is request after sending to the broker, before receiving
  response
* EndTxn goes through -- returns without doing anything because there
  are actually no records produced yet
* First request is REPEATED with sequence 0 -- because it because with 0
  it looks like a reset.
* We finally EndTxn

It can lead to duplicates to not Flush -- if there were outstanding
unsettled records, they could be duplicated.

The only safe thing is to flush, then end, no concurrent producing.