kgo: fixup the next-gen rebalancer as much as possible and disable it
After this commit, I will be removing the auto-opt in. Most of the code
here is valid. However, I don't believe Kafka's implementation is great.
There is no good solution for end users when an OffsetCommit fails with
STALE_MEMBER_EPOCH when the end user is leaving the group. This error
is extremely common because the broker bumps the member epoch A LOT.
It's not clear at all why the member epoch bumps so much. The epoch also
is not always recoverable.

KAFKA-19222: this issue points out that the KIP is documented incorrectly
KAFKA-19233: what looks to be a bug
KAFKA-19235: STALE_MEMBER_EPOCH is not really possible to be handled with OffsetCommit

Separately, the broker is NOT TOLERANT to ConsumerGroupRequest retries.
If clients always double the request, then clients regularly and
repeatedly get FENCED_LEADER_EPOCH. With 19233, this regularly stalls
consumers. With 19235, this regularly results in failed commits.

Until the broker is reliable in the face of repeated
ConsumerGroupHeartbeat requests, I lean to not opt in by default.

If somebody really wants to opt in, they can use a context with the
special value "opt_in_kafka_next_gen_balancer_beta", which I'll
mention only in the changelog.