kgo: make ErrFirstReadEOF not retryable
This isn't perfect, and may lead to some requests failing when they
shouldn't. If people report problems with this, a much more invasive
change will need to be made.

If SASL is misconfigured, brokers close connections immediately. EOF
errors should be retryable -- a connection getting cut is somewhat
common. BUT, if the first read errors, we'll pessimistically assume the
problem is SASL and not retry.

However, the downside is that if a broker is overloaded and closes a
valid connection immediately, or if a connection dies immediately, then
we may cause a request to fail when we would previously retry. Ideally
that does not happen much.